Relocated scrapping files under src
