+++ src/scrappers/source/proCyclingStats/raceStageResults.js
@@ -1,8 +1,8 @@
 import { Page } from "puppeteer-core";
-import { generateId } from "../../src/utils/idGenerator";
-import { renameKeys } from "../../src/utils/object";
-import { toCamelCase } from "../../src/utils/string";
-import { addTime, formatSeconds } from "../../src/utils/time";
+import { generateId } from "../../../utils/idGenerator";
+import { renameKeys } from "../../../utils/object";
+import { toCamelCase } from "../../../utils/string";
+import { addTime, formatSeconds } from "../../../utils/time";
 
 /**
  * @param {Array<Object>} table - The table to clean up.
@@ -109,7 +109,7 @@ function cleanUpStageTable(table, additionalValues) {
 /**
  *
  * @param {string} label - The label for the sprint.
- * @returns {Object<{location: string, distance: string}>} The parsed sprint label.
+ * @returns {{location: string, distance: string}} The parsed sprint label.
  */
 function sprint(label) {
   // Sprint -> "Sprint | Dozza (108.1 km)""
@@ -138,7 +138,7 @@ function sprint(label) {
 /**
  *
  * @param {string} label - The label for the climb.
- * @returns {Object<{category: string, location: string, distance: string}>} The parsed climb label.
+ * @returns {{category: string, location: string, distance: string}} The parsed climb label.
  */
 function climb(label) {
   // Stage Classification -> "KOM Sprint (3) Côte de San Luca (186.6 km)""
