+++ src/scrappers/source/proCyclingStats/raceStages.js
@@ -1,7 +1,7 @@
 import { Page } from "puppeteer-core";
-import { generateId } from "../../src/utils/idGenerator.js";
-import { formatDate } from "../../src/utils/string.js";
-import { logError } from "../../src/utils/logging.js";
+import { generateId } from "../../../utils/idGenerator.js";
+import { formatDate } from "../../../utils/string.js";
+import { logError } from "../../../utils/logging.js";
 
 /**
  * Cleans a record by extracting stage number and type, and generating IDs.
