+++ src/scrappers/source/proCyclingStats/races.js
@@ -1,8 +1,8 @@
 import { Page } from "puppeteer-core";
-import { generateId } from "../../src/utils/idGenerator";
-import { logError, logOut } from "../../src/utils/logging";
-import { formatDate } from "../../src/utils/string";
-import { buildUrl, urlSections } from "../../src/utils/url";
+import { generateId } from "../../../utils/idGenerator";
+import { logError, logOut } from "../../../utils/logging";
+import { formatDate } from "../../../utils/string";
+import { buildUrl, urlSections } from "../../../utils/url";
 
 /**
  * Collects World Tour races for a given year
